+++ post-formats/content-case-study.php
@@ -1,4 +1,4 @@
-
+<!-- content-case-study.php -->
 	<a class="m-1of3 t-1of3 d-1of3 pad-top" href="<?php echo esc_url( get_permalink() ); ?>">
 <?php // Grab the image to the case study ?>
 	<div class="case-study-icon">