+++ post-formats/content-listing.php
@@ -1,3 +1,5 @@
+<!-- content-listing.php -->
+
 <div class="m-all t-1of2 d-1of3">
 
 		<div class="listing-box cf">