+++ post-formats/content-serviceDropdown.php
@@ -1,3 +1,4 @@
+<!-- content-serviceDropdown.php -->
 
 <ul class="tabs d-1of2 m-all">
   <li class="active" rel="tab1"><h5 class="header-dark">Auction Services</h5></li>
@@ -9,35 +10,35 @@
 
 <div class="tab_container d-1of2 m-all">
 
-  <h3 class="d_active tab_drawer_heading" rel="tab1">Auction Services</h3>
+  <h3 class="d_active tab_drawer_heading" rel="tab1"><i class="fa fa-arrow-right" aria-hidden="true"></i>  Auction Services</h3>
   <div id="tab1" class="tab_content">
   <img src="<?php echo get_template_directory_uri(); ?>/library/images/icons/icon-auction-gavel.svg" alt="">
   <h4>Going, Going, SOLD.</h4>
     <p>Sell your property quickly and effectively. Put our experienced auction team to work for you!</p>
   </div>
 
-  <h3 class="tab_drawer_heading" rel="tab2">Occupier Representation</h3>
+  <h3 class="tab_drawer_heading" rel="tab2"><i class="fa fa-arrow-right" aria-hidden="true"></i>  Occupier Representation</h3>
   <div id="tab2" class="tab_content">
   <img src="<?php echo get_template_directory_uri(); ?>/library/images/icons/icon-lease.svg" alt="">
   <h4>Occupier Representation</h4>
     <p>At RESOURCE, you can be on top of the constantly changing market. We help you find your ideal location and customize it to your needs.</p>
   </div>
 
-  <h3 class="tab_drawer_heading" rel="tab3">Advisory and Consultation</h3>
+  <h3 class="tab_drawer_heading" rel="tab3"><i class="fa fa-arrow-right" aria-hidden="true"></i>  Advisory and Consultation</h3>
   <div id="tab3" class="tab_content">
   <img src="<?php echo get_template_directory_uri(); ?>/library/images/icons/icon-lease.svg" alt="">
   <h4>A Team that Gets You Places.</h4>
     <p>Get the advice, expertise and counsel to ensure you can make the best decisions for your business. If you want to find a new space to grow, evaluate your lease or find other solutions, we can help!</p>
   </div>
 
-  <h3 class="tab_drawer_heading" rel="tab4">Owner Representation</h3>
+  <h3 class="tab_drawer_heading" rel="tab4"><i class="fa fa-arrow-right" aria-hidden="true"></i>  Owner Representation</h3>
   <div id="tab4" class="tab_content">
   <img src="<?php echo get_template_directory_uri(); ?>/library/images/icons/icon-brokerage-sale.svg" alt="">
   <h4>Own Like a Pro.</h4>
     <p>RESOURCE helps you sell your property, buy a new property, find tenants for your building or manage the building you have. You have all the tools and people you need to get where you want to be.</p>
   </div>
 
-  <h3 class="tab_drawer_heading" rel="tab5">Development</h3>
+  <h3 class="tab_drawer_heading" rel="tab5"><i class="fa fa-arrow-right" aria-hidden="true"></i>  Development</h3>
   <div id="tab5" class="tab_content">
   <img src="<?php echo get_template_directory_uri(); ?>/library/images/icons/icon-land-escavator.svg" alt="">
   <h4>Build Something Great.</h4>